==> wk-3_python-project-2/2-data-aggregation.ipynb <==
# Data wrangling and datetimes

Dates and times are a special kind of data type. In this notebook, we will keep exploring the `orders` and `orderlines` datasets from Eniac and learn to deal with them.

import pandas as pd
df = pd.read_csv('../data/eniac/orderlines.csv')
df.head()

## Data exploration

df.info()

Two variables need to be modified: 

* `unit_price`: it is detected as an object but it has to be a float. Why is that happening? 
* `date`: needs to be transformed to a date time format. 

For the moment we will only focus on the `date`. Since it contains both the date and the time, we will transform the data type using the pandas method `to_datetime`: 

df['date'] = pd.to_datetime(df['date'])
df.info()

df.describe()

You can count the quantity of observations with `.value_counts()`.

df['sku'].value_counts()

### Exercise: check if the dates from the `orderlines` dataset match with the dates on the `orders` dataset.

orderlines = pd.read_csv('data/eniac/orderlines.csv', 
                         dtype={'id': int, 
                                'id_order': int, 
                                'product_id': int,
                                'sku': str, 
                                'unit_price': str, 
                                'date': str}, 
                         parse_dates=['date'])

orders = pd.read_csv('data/eniac/orders.csv', 
                     dtype={'order_id': int, 
                            'created_date': str, 
                            'total_paid': float,
                            'state': str}, 
                     parse_dates=['created_date'])


# print(f"{}")
print(f"The first and last date in the orderlines csv are {orderlines.date.max()} and {orderlines.date.min()} respectively")
print(f"The first and last date in the orders csv are {orders.created_date.max()} and {orders.created_date.min()} respectively")
print('\n')

print(f"There are {orderlines.date.nunique()} unique dates in the orderlines csv.")
print(f"There are {orders.created_date.nunique()} unique dates in the orders csv.")
print('\n')

dates_not_in_orders = orderlines[~orderlines.date.isin(orders.created_date)].shape[0]
dates_not_in_orderlines = orders[~orders.created_date.isin(orderlines.date)].shape[0]
print(f"{dates_not_in_orderlines} dates from orders do not exist in orderlines")
print(f"{dates_not_in_orders} dates from orderlines do not exist in orders")
print('\n')


print("If we round the timestamps to the minute then:")
dates_to_minute_not_in_orderlines = orders[~orders.created_date.dt.minute.isin(orderlines.date.dt.minute)].shape[0]
dates_to_minute_not_in_orders = orderlines[~orderlines.date.dt.minute.isin(orders.created_date.dt.minute)].shape[0]
print(f"{dates_to_minute_not_in_orderlines} dates from orders do not exist in orderlines")
print(f"{dates_to_minute_not_in_orders} dates from orderlines do not exist in orders")

print('\n')
print("This implies that it might take some time between orders being placed and the orderlines database being updated with the individual products")


merged_orders = orderlines.merge(
    orders, 
    how='left', 
    left_on='id_order', 
    right_on='order_id'
).assign(
    check_date = lambda x: x['date'] == x['created_date']
)

merged_orders[merged_orders.check_date == False].shape[0]

## Aggregating with pandas

* Grouping and aggregating is one of the main ways to explore data. The main tools to do that with pandas are:
    * [`pandas.DataFrame.groupby()`](https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.groupby.html).
    * [`pandas.DataFrame.agg()`](https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.agg.html?highlight=agg#pandas.DataFrame.agg) 
        * [`pandas.DataFrame.count()`](https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.count.html)

df.head()

How many id do we have on each sku?

df.groupby(['sku']).agg({'product_quantity':'count'})

Which is the total quantity acquired for each `sku`? And the median? And the mean?

df.groupby(['sku']).agg({'product_quantity':['count', 'sum','median','mean']})

`groupby` can also be combined with other pandas functions to analyse in more depth the datasets.

# WHY IS THE SKU COLUMN CALLED ALONG WITH PRODUCT QUANTITY?
df.groupby('sku')[['sku','product_quantity']].describe()

Now let's see how can I combine `.groupby()` and `.aggregate()` with `.sort_values()`.

I would like to see the top 10 most sold products on our data (total product quantity). 

(
df
    .groupby('sku')['product_quantity']
    .sum()
    .sort_values(ascending=False).head(10)
)

How can I sort values with multiple aggregated functions?

(
df
    .groupby('sku')
    .agg({'product_quantity':['sum','count','std','mean','median']})
    .sort_values(('product_quantity','mean'), ascending=False)
)

## Working with date time

Time to create the week day column. You will have to combine two functions, [`pandas.DataFrame.assign()`](https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.assign.html) from `pandas` and [`datetime.strftime()`](https://docs.python.org/3/library/datetime.html#strftime-and-strptime-behavior) from `datetime`.

First, let's talk about `.assign()`. It is a useful method for creating new columns:

df.assign(new_colum = 'hi! I am a new column!').head()

A new column often is the result of an operation between other columns in the dataframe:

df.assign(total_price = df['product_quantity'] * df['unit_price']).head()

In Pandas, we can extract strings containing elements like the day of the month, the day of the week, the hour of the day... from `datetime` objects. We do so by using `dt.strftime()` in combination with `strftime` syntax. Find the cheat sheet for this syntax [here](https://strftime.org/). 

You'll understand it better with an example. Let's create the weekday column:

df.assign(week_day = df['date'].dt.strftime('%A')).head()

Now we can find out which are the week day with more sells in quantity: 

(
df
    .assign(week_day = df['date'].dt.strftime('%A'))
    .groupby('week_day')
    .agg({'product_quantity':['sum','count']})
    .sort_values(('product_quantity','sum'), ascending=False)
)

Is there a way to find out the top 1 product (in product quantity) by each sku in each weekday?

(
df
    .assign(week_day = df['date'].dt.strftime('%A'))
    .groupby(['week_day','sku'])
    .agg({'product_quantity':'sum'})
    .sort_values('product_quantity', ascending=False)
    .reset_index()
    .groupby('week_day')
    .head(1)
)

# CHALLENGES

### 1. Which are the top 6 orders with the highest number of products sold?

(
    orderlines
    .groupby('id_order')
    .agg({'product_quantity':'sum'})
    .sort_values('product_quantity', ascending=False)
    .head(6)
)

### 2. Which are the top 6 orders for the 1 of July of 2017? 

Combining `.assign()` with `.dt.date` will help you extract the date of a datetime column and use this information to filter it. Do some googleing to find out how to use `.dt.date`.

Here is an example how to filter the first of January 2017.

# Example
df_date = df.assign(date = df['date'].dt.date).copy()
df_date[df_date['date'] == pd.to_datetime('2017-01-01')]

# I am assuming that by 'top 6 orders' they mean the orders with the largest product_quantity
(
orderlines
    .loc[orderlines.date.dt.date == pd.Timestamp('2017-07-01')]
    .groupby('id_order')
    .agg({'product_quantity':'sum'})
    .sort_values('product_quantity', ascending=False)
    .head(6)
)

### 3. Which is the month with the highest number of units sold? Remember to look at the documentation of [`datetime.strftime()`](https://docs.python.org/3/library/datetime.html#strftime-and-strptime-behavior) 

(
orderlines
    .assign(month = orderlines.date.dt.strftime('%B')) # dt.month)
    .groupby(['month'])
    .agg({'product_quantity':'sum'})
    .sort_values('product_quantity', ascending=False)
    .head(6)
)

The results show enormous differences between some months. Why do you think this is happening? Do we only have one year of data? In case you have multiple years, filter for only 2017 to find out the best month for that year.

ATTENTION: the output of the function `.dt.strftime()` is a string!

(
orderlines
    .loc[orderlines.date.dt.year == 2017]
    .assign(month = orderlines['date'].dt.strftime('%B'))
    .groupby(['month'])
    .agg({'product_quantity':'sum'})
    .sort_values('product_quantity', ascending=False)
    .head(6)
)

# Option 2
(
orderlines
    .assign(month = orderlines['date'].dt.strftime('%B'),
           year = orderlines['date'].dt.strftime('%Y'))
    .query("year == '2017'")
    .groupby(['month'])
    .agg({'product_quantity':'sum'})
    .sort_values('product_quantity', ascending=False)
    .head(6)
)

### 4. Find the day of the week with the highest amount of products sold by each MONTH. See an idea of the expected output. Take into account the year problem we mentioned before.

pd.DataFrame({
    'month':['November','January','July','...'],
    'week_day':['Friday','Tuesday','Monday','...'],
    'product_quantity':[10000, 9800, 9700, 0]
})

(
orderlines
    .assign(year = orderlines['date'].dt.strftime('%Y'),
            month = df.date.dt.strftime('%B'), 
            week_day = df.date.dt.strftime('%A')) # dt.month)
    .query("year == '2017'")
    .groupby(['month', 'week_day'])
    .agg({'product_quantity': 'sum'})
    .sort_values('product_quantity', ascending=False)
    .reset_index()
    .groupby('month')
    .head(1)
)
