==> wk-3_python-project-2/3-data-cleaning-and-merging.ipynb <==
# Data cleaning and merging dataframes

## Loading multiple datasets

### Classical way

Reading file by file

import pandas as pd

path = '../data/eniac/'
orderlines = pd.read_csv(path + 'orderlines.csv')
orders = pd.read_csv(path + 'orders.csv')
brands = pd.read_csv(path + 'brands.csv')
products = pd.read_csv(path + 'products.csv')

df_list = [orderlines, orders, brands, products]
files = ['orderlines','orders','brands','products']

### Another way

Using `os` and a loop to read all the files from a directory. It's also possible to read only files with a certain extension (like `.csv`):

import pandas as pd
import os
path = '../data/eniac/'
path, dirs, files = next(os.walk(path))
print(files)

# remove non-csv files
for file in files:
    if not file.endswith("csv"):
        files.remove(file)

# create empty list 
df_list = []

# append datasets to the list
for file in files:
    temp_df = pd.read_csv(path + file, sep=',')
    df_list.append(temp_df)

# show results
orders, brands, orderlines, products = df_list[0], df_list[1], df_list[2], df_list[3]

## Data quality

### Missing values

def check_missing_values(data): 
    print('Missing values:' + '\n' + str(data.isna().sum()))

for i in range(len(df_list)): 
    print(f'File name: {files[i]}')
    check_missing_values(df_list[i])
    print('\n')

### Duplicates

def check_duplicates(data): 
    print('Duplicated rows: ', data.duplicated().sum())

for i in range(len(df_list)): 
    print(f'File name: {files[i]}')
    check_missing_values(df_list[i])
    check_duplicates(df_list[i])
    print('\n')

Based on our initial exploration we need to find a way to deal with our missing values. 

Do we have other problems like data inconsistency? 

* Example: problems with the `price` column in the `orderlines` dataframe.

### Data inconsistency

for i in range(len(df_list)): 
    
    print(f'File name: {files[i]}')
    print('-----------------------------------')
    check_missing_values(df_list[i])
    print('-----------------------------------')
    check_duplicates(df_list[i])
    print('-----------------------------------')
    df_list[i].info()
    print('\n')

Some aspects that do not make a lot of sense:

* **products**: 
    * `price` and `promo_price` are loaded as objects. They should be floats. 
    * Quite a lot of missing values: 
        * `description`: 7 missing values. Maybe that could be inferred from the product name?
        * `price`: the missing values could be filled from the `orderlines` dataset. But first we will need to clean it. 

    
* **orders**: 
    * `created_date` should have a date data type. Then, it would be a good quality check to see if the created date for `orders` matches with the created dates for `orderlines`. 
    
    
* **orderlines**: 
    * `unit_price` has to be a float, something wrong there. 
    * `date` has to be transformed to a date data type. Then, as we said, check that it with matches with the `orders` dataset. 
    
    
* **brands**: looks fine. 


Where do we have to start? Since `orders` and `orderlines` seem very crucial to the analysis, we will start by checking that they match.

## Cleaning orders

'''
Products

drop 
promo price - hard to clean
in stock
type

'''

# change date datatype
orders['created_date'] = pd.to_datetime(orders['created_date'])

Some data I could extract from here to do a data quality check with `orderlines`: 

* the sum of `total_paid` on both datasets should be the same
* the initial and last date should be the same

min(orders['created_date'])

max(orders['created_date'])

sum(orders['total_paid'])

I forgot to deal with missing values on our orders dataset!

orders.isna().sum()

# let's filter them out
orders.loc[orders['total_paid'].isna()]

These orders have a state of "Pending", so I can exclude them from our dataset. We are interested on the completed ones.

orders.dropna(inplace=True)

Now my dataset is clean. And the total paid is: 

orders['total_paid'].sum()

## Cleaning orderlines

First let's transform our date time. 

orderlines['date'] = pd.to_datetime(orderlines['date'])

min(orderlines['date'])
# orders: Timestamp('2017-01-01 00:07:19')

max(orderlines['date'])
# orders: Timestamp('2018-03-14 13:58:36')

Yes, it seems it match. Now let's test our total_paid. 

To get this value, we will have to calculate a new column, total price for each row. It would be `product_quantity` * `unit_price`. I can not multiply a two different types of data. So I will have to transform the unit price to a numeric one. 

orderlines['product_quantity'] * pd.to_numeric(orderlines['unit_price'])

While trying to transform this column to numeric an error appears. From the error message, "Unable to parse string "1.137.99" at position 6", it seems that our dataset has some problems with the thousands separators: they were encoded as dots, and python only admits one dot per number: the _decimal_ separator!

There are a lot of different ways to solve this problem. The first thing I will do is to count how many dots appear for each number using string operations. If price has more than one dot, I will consider it corrupted. 

# we create a copy of the dataset
ol_temp = orderlines.copy()

# create a new column with the amount of dots in the unit_price column
# The \ is used before the . here because usually . is used to represent all characters. Without escaping the . this code might count all characters.
ol_temp['dots'] = orderlines['unit_price'].str.count('\.')

# show the rows with more than one dot
ol_temp.query('dots > 1')

Our theory about the thousands separators is confirmed. How can we solve this problem? 

My approach would be to remove all the dots for all the unit_prices, and then add a dot before the last 2 digits to all the rows. Then I will transform it into numeric values.

orderlines = orderlines.assign(unit_price_nd = orderlines['unit_price'].str.replace('\.','', regex=True))
orderlines.head()

orderlines['digits'] = orderlines['unit_price_nd'].str[-2:]
orderlines['integers'] = orderlines['unit_price_nd'].str[:-2]
orderlines.head()

orderlines['new_unit_price'] = orderlines['integers'] + '.' + orderlines['digits']
orderlines.head()

# last step is to transform it to numeric
orderlines['unit_price'] = pd.to_numeric(orderlines['new_unit_price'])
orderlines.info()

Now it is time to see if it is the same than the orders files. 

orderlines.drop(['unit_price_nd','digits','integers','new_unit_price'], axis=1, inplace=True)

orderlines['total_price'] = orderlines['product_quantity'] * orderlines['unit_price']
sum(orderlines['total_price'])
# 129159615.07000001

''' 
Alternative solutions: 

df['unit_price'] = df['unit_price'].str.split('.',1).apply(lambda x : x[0]+ '' + x[1])

df['unit_price'] = [ i.replace(".","") for i in df['unit_price'] ] 
df['unit_price'] = pd.to_numeric(df['unit_price'])
df['unit_price'] = df['unit_price']/100

df['unit_price'] = df['unit_price'].str.rsplit('.',1).apply(lambda x : x[0]+ ',' + x[1]).str.replace('.','',regex=True).str.replace(',','.',regex=True)
'''

What!? After all this work the values do not match! How can we figure it out where is the problem? 

## Matching `orders` and `orderlines`

It is possible that some orders exist in one dataset but not in the other one. This would be a potential source for this price mismatch. Let's solve it!

We first create a new column in the `orderlines` dataset using `assign`. We also use `isin()` to create a boolen value (True/False) that checks whether the `id_order` is present in the `orders` dataset:

(orderlines.
 assign(check_orders = orderlines.id_order.isin(orders.order_id))
)

Then, using `.query` we select rows where the value in this new column is `False`:

(orderlines.
 assign(check_orders = orderlines.id_order.isin(orders.order_id)).
 query("check_orders==False")
)

It looks like 240 rows in `orderlines` come from orders not present in the `orders` dataset. This is quite inconsistent, since the `orders` dataset should be the one and only source of truth for orders: if an order is not there, it should not exist. We will try to find out why that happened, but for now, let's just remove those "ghost" orders:

orderlines = (
    orderlines
        .assign(check_orders = orderlines.id_order.isin(orders.order_id))
        .query("check_orders==True")
)

Now let's look at this problem in the opposite direction: are there orders in the `orders` dataset not prsent in `orderlines`?

(
    orders
        .assign(check_orders = orders.order_id.isin(orderlines.id_order))
        .query("check_orders==False")
)

There are more than 22000 orders in the `orders` dataset that are not present on the `orderlines` dataset!!! We can try to find out why by looking at the state of these orders:

(
    orders
        .assign(check_orders = orders.order_id.isin(orderlines.id_order))
        .query("check_orders==False")
        .state.value_counts()
)

It looks like most of them are orders that were not fully completed: the products were left in the shopping basket or the order was "placed" but maybe not paid (hence the state "Place Order". Some of them were "Completed", though. 

This will require further research, and we might have to come back to these orders if we have to explore consumer behaviour (e.g. why are orders left in the shopping basket?), but for now, for the sake of data quality, let's drop all of these unmatched orders:

orders = (
    orders
        .assign(check_orders = orders.order_id.isin(orderlines.id_order))
        .query("check_orders==True")
)

Let's now check again if the total paid matches:

orders.total_paid.sum()

orderlines.total_price.sum()

STILL NOT MATCHING!!! This is outrageous. Let's keep exploring.

## Solving the price mismatch

Let's merge both datasets and compare, order by order, the `total_price`. We will call this new merged dataset `orders_info`.

*Note: To merge `orderlines` with `orders`, we have to group and aggregate `orderlines`!

# Left join is acceptable here because we have already removed rows that do not match in both dataframes.
# Otherwise an inner join would be more suitable.
orders_info = (
    orderlines
        .groupby('id_order')
        .agg({'total_price':'sum'})
        .merge(orders, how='left', left_on='id_order', right_on='order_id')
)
orders_info

Now that the `total` from both datasets is in the same dataframe, we can create a new column with the difference:

orders_info['price_difference'] = orders_info['total_paid'] - orders_info['total_price']
orders_info

orders_info.price_difference.describe()

Looks like the maximum and minimun price differences are huge: some orders are really corrupted. But we also see from the mean and the quartiles that the price difference is neglegible for most orders.

## Challenge: Remove outliers

Decide on a criteria for removing orders whenever you cannot trust the price difference between `orders` and `orderlines`. 

orders_info.price_difference = round(orders_info.price_difference, 2)

oi = orders_info.price_difference.unique().tolist().sort()
print(oi)

Save the data once you are sure you can trust it!

orderlines.to_csv(path + 'clean/orderlines_cl.csv', index=False)
orders.to_csv(path + 'clean/orders_cl.csv', index=False)

## Challenge: Cleaning products

Now it is time to clean the products dataset. Let's do a quick review of the things where wrong: 

print(products.head())
print('\n')
print('\n')
print(products.info())
print('\n')
print('\n')
check_duplicates(products)
print('\n')
print('\n')
check_missing_values(products)

Looking at this overview, we can see that there are different things that has to be changed: 

* Data types: 
    * price should be a float
    * promo price should be a float
* Duplicated rows. They have to be removed. 
    * To accomplish this step you can use the method `pd.DataFrame.drop_duplicates()`. Be sure you drop all the duplicates based on the column **sku**, as it is the one you will use to merge with orderlines. 
* Missing values: 
    * Description may be can be infered by the name
    * price. Is there a way we can extract the information from another table?
    * type. Do we need this column for our analysis?
    
This task can be accomoplished by using all the methods you already know.

### Start of the challenge

# Drop the type column and duplicate rows based on sku value
temp = products.copy()
temp.drop('type', axis=1, inplace=True)

temp.drop_duplicates(subset='sku', inplace=True)

# Check for products without descriptions
names_of_products_without_descriptions = temp[temp.desc.isna()].name.tolist()
names_of_products_without_descriptions

# Add descriptions
missing_descriptions = [
    '2TB Mac hard drive and Nas',
    'Apple keyboard for iPad 9.7',
    'NAS server with 10GB RAM',
    'Ethernet adapter for Macbook 12',
    'Luxury power bank combined with powder, 2 mirrors - normal and 3x magnification, Illuminated under mirror with LED, Low weight and compact dimensions',
    'Battery capacity: 20,000 mAh; ultra-stable: outer shell made of durable synthetic rubber (military standard, withstands drops from up to 2 metres) ; protection: dust and splash proof: military standard iP54; battery level indicator and super fast charging; USB port can be connected to charger and other devices',
    'Smart thermostat designed to provide automatic time and temperature control of heating systems in homes and apartments. '
]
for i in range(len(names_of_products_without_descriptions)):
    temp.loc[temp.name == names_of_products_without_descriptions[i], 'desc'] = missing_descriptions[i]
    
for prod_name in names_of_products_without_descriptions:
    print(temp[temp.name == prod_name].desc)

# Clean the price and promo_price columns

def start_pipeline(df):
    '''Make a copy of the pipeline to prevent corrupting the original data'''
    return df.copy()

def remove_duplicates(df, col):
    '''Remove duplicate rows'''
    return df[~df[col].isna()]

def clean_prices(df):
        df.price = df.price.apply(lambda x : x +'.00' if x.count('.') == 0 else x)
        df.price = df.price.apply(lambda x: x  + '0' if x[-2]=='.' else x)
        df.price = df.price.apply(lambda x: str(float(x.replace('.',''))/1000) if ( (x[-4]=='.') & (x.count('.')==2)) else x)
        df.price = df.price.apply(lambda x: str(float(x.replace('.',''))/10000) if ( (x[-4]=='.') & (x.count('.')==1)) else x)
        df.price = df.price.astype(float)
        df.price = df.price.round(decimals = 2)
        return df

def split_and_join_promo_prices(df):
    '''Remove the decimal points from the promo_price strings and append .00 to the string'''
    return df.assign(promo_price=df.promo_price
                     .str.split('.')
                     .str.join('')
                     .apply(lambda x: x+'.00')
                    )

def _insert_decimal_at_string_position(s, pos):
    '''Insert a decimal point at a given position in a string'''
    s = s.split('.')
    s = s[0] + s[1]
    s = s[:pos]+'.'+s[pos:]
    return s 


def _insert_decimal_in_promo_price(row):
    '''
    If the euro value of the price is equal to the euro value of the promo_price, 
    set the promo_price equal to price and return it.
    This is because some promo_prices are slightly larger than their equivalent prices, e.g. 12.95 - 12.99
    
    Otherwise, keep moving the decimal point towards the start of the string until the promo_price is lower than the price.
    Then transform the string to a float and round it to two decimal places.
    '''
    
    decimal_position = -2
    while float(row.promo_price) > row.price:
        if round(float(row.promo_price), 0) == round(row.price, 0):
            row.promo_price = row.price
            return row.promo_price
        else:
            row.promo_price = _insert_decimal_at_string_position(row.promo_price, decimal_position)
            decimal_position += -1
    
    return round(float(row.promo_price), 2)

def transform_promo_price_to_floats(df):
    df.promo_price = [_insert_decimal_in_promo_price(row) for index, row in df.iterrows()]
    return df

temp = (products
        .pipe(start_pipeline)
        .pipe(remove_duplicates,col='price')
        .pipe(clean_prices)
        .pipe(split_and_join_promo_prices)
        .pipe(transform_promo_price_to_floats)
)

print(temp[temp.promo_price > temp.price])
print('----------------------------------------------------')
print('\n')

temp.head()

# Check values

print('\n')
print(temp[temp.sku=='SYN0058'])

print('\n')
print(temp[temp.sku=='CRU0015-2'])

print(temp[temp.price == 59.99][['price', 'promo_price']])     # 59.99 - 499.899
print('\n')
print(temp[temp.price == 2159.00][['price', 'promo_price']])    # 2159 - 1769
print('\n')
print(temp[temp.price == 2399.00][['price', 'promo_price']])    # 2399 - 23.989.896
print('\n')
print(temp[temp.price == 1319.00][['price', 'promo_price']])    # 1319 - 1279.99
print('\n')
print(temp[temp.price == 79.99][['price', 'promo_price']])    # 79.99 - 75.988
print('\n')
print(temp[temp.price == 107.99][['price', 'promo_price']])    # 107.99 - 559.903

print('\n')
print(temp[temp.price == 24.950][['price', 'promo_price']])    # 24.95 - 24.99

temp[temp.promo_price==None]

products.head()

## Brands

The brands csv looks fine, so we can work together with it. 

brands.to_csv(path + 'clean/brands.csv', index=False)

t = temp.head(5)
t2 = temp.head(10)

t3 = t2[~t2.index.isin(t.index)]

print(t3)
