==> wk-3_python-project-2/5-category-creation.ipynb <==
import pandas as pd
products = pd.read_csv('../data/eniac/products_clean.csv')

import re

# Introduction to Regex

## GOAL

Introduce to the library `re` (regexp) and show the main functions and how filter text based on regular expressions. 

## DESCRIPTION

In this workshop, the following functions will be reviewed: 

* `findall()`
* `search()`
* `split()`
* `sub()`
* `span()`
* `string()`
* `group()`

Metacharacters: ` . ^ $ * + ? { } [ ] \ | ( )`

Special Sequences: `\A` `\b` `\d` `\s`

And how to compile the regex expressions to reuse it. 

More information on that [link](https://www.w3schools.com/python/python_regex.asp).

products.sample()

# extract a specific description
prod_descr = products.query('sku == "DLK0139"')['desc'].values[0]
prod_descr

### `findall`

Returns a list containing all matches

# return all occurrences appearing on a string
re.findall('a', prod_descr)

### `search`

Returns a Match object if there is a match anywhere in the string. If there is more than one match, only the first occurrence of the match will be returned.

The match objects have the following methods: 
- `.span()` returns a tuple containing the start-, and end positions of the match.
- `.string` returns the string passed into the function
- `.group()` returns the part of the string where there was a match

prod_descr

match_obj = re.search('video', prod_descr)

match_obj.string

match_obj.group()

match_obj.span()

### `split`

Returns a list where the string has been split at each match

prod_descr.split(' and ') # and is removed from the list

### `sub`

Replaces one or many matches with a string

dark_descr = re.sub("camera", "pool", prod_descr)
print(dark_descr)

### METACHARACTERS


Some characters are special metacharacters, and don’t match themselves. Instead, they signal that some out-of-the-ordinary thing should be matched, or they affect other portions of the RE by repeating them or changing their meaning.

` . ^ $ * + ? { } [ ] \ | ( )`

 #### `[]` means set of characters:
 
 - `[abc]` will match any of the characters a, b, or c
 - `[a-c]` will do the same
 - `[a-z]` will match any lowercase letter

alphanumeric = "4298fsfsv012rvv21v9"

re.findall(r"[a-z]", alphanumeric)

`\` Can help us to scape special characters 

alphanumeric_with_special = alphanumeric + "[a-z]"
print(alphanumeric_with_special)
# CHALLENGE: use \ to escape the square brakets
re.findall(r"\[a-z]", alphanumeric_with_special)

#### Some special sequences:

- `\A`- Returns a match if the specified characters are at the beginning of the string
- `\b` - Returns a match where the specified characters are at the beginning or at the end of a word
- `\d` - 	Returns a match where the string contains digits (numbers from 0-9) (`\D` for where the string DOES NOT contain digits)
- `\s`- Returns a match where the string contains a white space character (`\S` for where the string DOES NOT contain a white space)

prod_descr

# find all possible numbers
re.findall(r"\d", prod_descr)

### `.`	Any character (except newline character)

re.findall(r'..c', prod_descr)

### `+` One or more occurrences

re.findall(r'e+', prod_descr)

re.sub("e+", "__", prod_descr)

### `{}`- Exactly the specified number of occurrences

re.findall(r"e{2}", prod_descr)

re.sub(r"e{2}", "__", prod_descr)

### `^` Starts with

re.findall(r"^F", prod_descr)

#### How to apply it on the whole dataframe?

products.loc[products['name'].str.contains(r'^Fit')]

Learn more how to apply regexp and pandas: 

* https://kanoki.org/2019/11/12/how-to-use-regex-in-pandas/

### `*`	Zero or more occurences

similar_words = ["hey", "hay", "how", "h i j k", "h", "ha", "oops"]

# use "." to return all words starting with "h"
for word in similar_words:
    print(re.findall("h.*", word))

print(prod_descr)
re.findall("vi*\S", prod_descr)

# Another way to show
re.findall("vi*\w+", prod_descr)
# \w: Returns a match where the string contains any word characters 
#    (characters from a to Z, digits from 0-9, and the underscore _ character)
# +: One or more occurrences

### Examples into dataframes

# I would like to filter all the names that contain body
(
products
    .loc[products['name'].str.contains(r'(body|Body)')]
    .sort_values('name').head(5))

# CHALLENGE: how can you reduce the previous regexp expression?
(
products
    .loc[products['name'].str.contains(r'(b|B)ody')]
    .sort_values('name').head(5))

### Compile regular expressions

# the last will be the first ones
regexp_dict = {
    'ipod':'^.{0,7}apple ipod',
    'case':'(case|funda|housing|casing|folder)',
    'cable':'cable|connector|Lightning to USB|Wall socket|power strip',
    'battery':'battery',
    'headset':'(headset|headphones)',
    'mouse':'(mouse|trackpad)',
    'stand':'(stand|support)',
    'protect':'(protect|cover|sleeve|Screensaver|shell)',
    'watch':'(^.{0,6}apple watch|smartwatch|smart watch)',
    'camera':'camera',
    'refurbished':'(refurbished|reconditioned|like new)',
    'strap':'strap|armband|belt|bracelet'
}

temp = products.copy().assign(category = 'unknown')

import numpy as np

for label, val in regexp_dict.items(): 
    regexp = re.compile(val, flags=re.IGNORECASE)
    temp = (
    temp
        .assign(
            category = lambda x: np.where(
                (x['desc'].str.contains(regexp, regex=True)) &
                (x['category'] == 'unknown'), label, x['category'])))

temp['category'].value_counts()
